requirements.txt
fastapi
uvicorn
pydantic
sqlmodel
sqlalchemy
python-dotenv
pytest
requests
httpx
